There are three ways to determine what to filter profile groups by:
1. Forced profile ID from custom fields
2. The selected ID they pick from the dropdown
3. The default "All Profiles" option.

Option 3 is where the problem was. It would query based on all groups coming back from the API rather than the groups offered in the dropdown selection. This new commit will allow you customize the dropdown array of groups and the final query to the API to get "All Profiles" will reflect whats in the dropdown.