Update to PHP-CS-Fixer v3.0.0
Move the installing of PHP-CS-Fixer from outside the project composer.json file to the recommended install method under /tools and install only when the environment is local. This allows it to be independent of the PHP site's composer.json if there were conflicts for the packages.

Update the .php-cs-fixer.php rules to match the Laravel Shift v3 and PHP8 config
https://gist.github.com/laravel-shift/cab527923ed2a109dda047b97d53c200#gistcomment-3755709

Update the git precommit hook to match the output for when there is a change needed to run `make phplint` instead of the dry run